tests: Set up dummy AWS region for any offline test boto3 clients
Offline tests use an arbitrary region to avoid any actual talking to the
infrastructure. Infrastructure tests require the actual region. And
since the region needs to be set up at import time I had to move some
infrastructure tests to separate files.

Because pylint wants imports from `backend` to all be grouped I could
not use `AWS_DEFAULT_REGION_KEY` in some places.
