+++ tests/test_notify_slack.py
@@ -1,32 +1,16 @@
 from datetime import datetime, timezone
 from http import HTTPStatus
-from json import dumps, load
+from json import dumps
 from logging import getLogger
 from os import environ
 from unittest.mock import MagicMock, patch
 
-from mypy_boto3_events import EventBridgeClient
-from mypy_boto3_lambda import LambdaClient
 from mypy_boto3_sns.type_defs import MessageAttributeValueTypeDef
-from pytest import mark
 from pytest_subtests import SubTests
 
 from backend.api_keys import EVENT_KEY
-from backend.api_responses import STATUS_CODE_KEY
+from backend.aws_keys import AWS_DEFAULT_REGION_KEY
 from backend.aws_message_attributes import DATA_TYPE_STRING
-from backend.notify_status_update.task import (
-    EVENT_DETAIL_KEY,
-    MESSAGE_ATTRIBUTE_DATASET_KEY,
-    MESSAGE_ATTRIBUTE_STATUS_KEY,
-    SLACK_URL_ENV_NAME,
-    STEP_FUNCTION_ARN_KEY,
-    STEP_FUNCTION_STARTDATE_KEY,
-    STEP_FUNCTION_STOPDATE_KEY,
-    WEBHOOK_MESSAGE_BLOCKS_KEY,
-    lambda_handler,
-    publish_sns_message,
-)
-from backend.resources import ResourceName
 from backend.step_function import Outcome
 from backend.step_function_keys import (
     ASSET_UPLOAD_KEY,
@@ -48,9 +32,24 @@ from backend.step_function_keys import (
     VERSION_ID_KEY,
 )
 
+from .aws_profile_utils import any_region_name
 from .aws_utils import any_arn_formatted_string, any_lambda_context, any_s3_url
 from .general_generators import any_https_url
 from .stac_generators import any_dataset_id, any_dataset_prefix, any_dataset_version_id
+
+with patch.dict(environ, {AWS_DEFAULT_REGION_KEY: any_region_name()}, clear=True):
+    from backend.notify_status_update.task import (
+        EVENT_DETAIL_KEY,
+        MESSAGE_ATTRIBUTE_DATASET_KEY,
+        MESSAGE_ATTRIBUTE_STATUS_KEY,
+        SLACK_URL_ENV_NAME,
+        STEP_FUNCTION_ARN_KEY,
+        STEP_FUNCTION_STARTDATE_KEY,
+        STEP_FUNCTION_STOPDATE_KEY,
+        WEBHOOK_MESSAGE_BLOCKS_KEY,
+        lambda_handler,
+        publish_sns_message,
+    )
 
 STEP_FUNCTION_START_MILLISECOND_TIMESTAMP = round(
     datetime(
@@ -243,36 +242,3 @@ def should_publish_sns_message(get_param_mock: MagicMock) -> None:
 
     # Then
     assert sns_client_mock.call_args[1] == expected_sns_call
-
-
-@mark.infrastructure
-def should_launch_notify_slack_endpoint_lambda_function(
-    lambda_client: LambdaClient, events_client: EventBridgeClient
-) -> None:
-
-    notify_status_lambda_arn = events_client.list_targets_by_rule(
-        Rule=ResourceName.CLOUDWATCH_RULE_NAME.value
-    )["Targets"][0]["Arn"]
-
-    # When
-    body = {
-        EVENT_DETAIL_KEY: {
-            STATUS_KEY: JOB_STATUS_FAILED,
-            INPUT_KEY: dumps(
-                {
-                    DATASET_ID_KEY: any_dataset_id(),
-                    DATASET_PREFIX_KEY: any_dataset_prefix(),
-                }
-            ),
-        },
-        OUTPUT_KEY: None,
-    }
-
-    resp = load(
-        lambda_client.invoke(
-            FunctionName=notify_status_lambda_arn,
-            Payload=dumps(body).encode(),
-        )["Payload"]
-    )
-
-    assert resp.get(STATUS_CODE_KEY) == HTTPStatus.OK, resp
